feat: healthResponseDto 생성

+++ src/main/java/com/ewha/pudong/repository/HealthRepository.java
@@ -3,5 +3,7 @@ package com.ewha.pudong.repository;
 import com.ewha.pudong.domain.Health;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HealthRepository extends JpaRepository<Health, Integer> {
+
+public interface HealthRepository extends JpaRepository<Health,Long> {
+
 }
